+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - refactor: use constants for log messages target.
 - fix(src/utils/extensions.ts): caching of the list of available Quarto extensions.
 - fix(src/utils/extensions.ts): update cache expiration time for extensions list and display in log as ISO string.
+- chore: use webpack to bundle the extension.
 
 ## 0.6.0 (2025-01-24)