+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.7.0 (unreleased)
+## 0.7.0 (2025-01-30)
 
 - refactor(src/utils/network.ts): internal logging.
 - refactor(src/utils/extensions.ts): externalise Quick Pick UI tools.